refactor(frontend): split OrdererDetailPage and PartnerDetailPage into sub-components
OrdererDetailPage (1,154 -> 267 lines parent + 5 sub-components):
- AnnouncementCard: compact grid-style announcement card
- AnnouncementConditionsPanel: search/sort/filter side panel
- OrdererDetailHeader: header with category, name, stats
- OrdererInfoTab: basic info with accordions for details/stats/departments
- OrdererAnnouncementsTab: announcement card grid with pagination
- Shared types and constants

PartnerDetailPage (1,135 -> 318 lines parent + 2 sub-components):
- ProjectConditionsPanel: search/sort/filter side panel for projects
- PartnerDetailHeader: header with name, number, project count
- Shared types and constants

Pure structural refactor - no behavior changes. TypeScript strict mode
compilation verified with zero errors.

Partial fix for #90
