+++ judgesystem/frontend/app/src/pages/PartnerDetailPage.tsx
@@ -2,809 +2,29 @@ import { useState, useMemo, useCallback, useEffect } from 'react';
 import { useParams, useNavigate, useLocation } from 'react-router-dom';
 import {
   Box,
-  Button,
   Typography,
   Paper,
-  TextField,
-  InputAdornment,
   Tabs,
   Tab,
-  IconButton,
 } from '@mui/material';
-import {
-  Search as SearchIcon,
-  ArrowBack as ArrowBackIcon,
-  SwapVert as SortIcon,
-  FilterAlt as FilterIcon,
-  Close as CloseIcon,
-} from '@mui/icons-material';
-import { bidTypeConfig } from '../data';
-import { colors, pageStyles, fontSizes, iconStyles, borderRadius, rightPanelColors } from '../constants/styles';
-import { workStatusConfig } from '../constants/workStatus';
-import { evaluationStatusConfig } from '../constants/status';
-import { priorityLabels, priorityColors } from '../constants/priority';
-import { categories } from '../constants/categories';
-import { bidTypes } from '../constants/bidType';
-import { prefecturesByRegion } from '../constants/prefectures';
-import { organizationGroupsByRegion } from '../constants/organizations';
-import { NotFoundView, FloatingBackButton, ScrollToTopButton, FilterButton } from '../components/common';
+import { colors, pageStyles, fontSizes, borderRadius } from '../constants/styles';
+import { NotFoundView, FloatingBackButton, ScrollToTopButton } from '../components/common';
 import { RightSidePanel } from '../components/layout';
 import { PartnerBasicInfo, PartnerQualifications, PartnerHistory } from '../components/partner';
 import { useSidebar } from '../contexts/SidebarContext';
 import type { PastProject, PartnerDetail } from '../types/partner';
-import type { BidType } from '../types/announcement';
 import type { EvaluationStatus, WorkStatus, CompanyPriority } from '../types';
 import { getApiUrl } from '../config/api';
-
-// ソートオプション
-type SortOption =
-  | 'deadline_asc' | 'deadline_desc'
-  | 'evaluationStatus_asc' | 'evaluationStatus_desc'
-  | 'workStatus_asc' | 'workStatus_desc'
-  | 'priority_asc' | 'priority_desc'
-  | 'prefecture_asc' | 'prefecture_desc'
-  | 'evaluatedAt_asc' | 'evaluatedAt_desc';
-
-// ソートフィールド定義
-const SORT_FIELDS = [
-  { field: 'deadline', label: '締切', ascLabel: '近い', descLabel: '遠い' },
-  { field: 'evaluationStatus', label: '参加可否', ascLabel: '可能→不可', descLabel: '不可→可能' },
-  { field: 'workStatus', label: '着手状況', ascLabel: '未着手→完了', descLabel: '完了→未着手' },
-  { field: 'priority', label: '優先度', ascLabel: '高→低', descLabel: '低→高' },
-  { field: 'prefecture', label: '都道府県', ascLabel: 'あ→わ', descLabel: 'わ→あ' },
-  { field: 'evaluatedAt', label: '判定日', ascLabel: '古い→新しい', descLabel: '新しい→古い' },
-] as const;
-
-// フィルタータブ定義
-const FILTER_TABS = [
-  { id: 'evaluationStatus', label: '参加可否' },
-  { id: 'priority', label: '優先度' },
-  { id: 'workStatus', label: '着手状況' },
-  { id: 'bidType', label: '入札方式' },
-  { id: 'category', label: '種別' },
-  { id: 'prefecture', label: '都道府県' },
-  { id: 'organization', label: '発注機関' },
-] as const;
-
-// フィルター状態の型
-interface ProjectFilterState {
-  workStatuses: WorkStatus[];
-  evaluationStatuses: EvaluationStatus[];
-  priorities: CompanyPriority[];
-  bidTypes: string[];
-  categories: string[];
-  prefectures: string[];
-  organizations: string[];
-}
-
-// WorkStatus, EvaluationStatus, Priority の配列
-const WORK_STATUS_OPTIONS: WorkStatus[] = ['not_started', 'in_progress', 'completed'];
-const EVALUATION_STATUS_OPTIONS: EvaluationStatus[] = ['all_met', 'other_only_unmet', 'unmet'];
-const PRIORITY_OPTIONS: CompanyPriority[] = [1, 2, 3, 4, 5];
-
-// WorkStatusの順序（ソート用）
-const WORK_STATUS_ORDER: Record<WorkStatus, number> = {
-  not_started: 0,
-  in_progress: 1,
-  completed: 2,
-};
-
-// EvaluationStatusの順序（ソート用：all_met→other_only_unmet→unmet）
-const EVALUATION_STATUS_ORDER: Record<EvaluationStatus, number> = {
-  all_met: 0,
-  other_only_unmet: 1,
-  unmet: 2,
-};
-
-// 案件用表示条件パネル
-interface ProjectConditionsPanelProps {
-  searchQuery: string;
-  onSearchChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  sortOption: SortOption | null;
-  onSortChange: (option: SortOption | null) => void;
-  filters: ProjectFilterState;
-  onFilterChange: (filters: ProjectFilterState) => void;
-  onClearAll: () => void;
-  activeTab?: 'sort' | 'filter';
-  onTabChange?: (tab: 'sort' | 'filter') => void;
-  projects: PastProject[];
-}
-
-function ProjectConditionsPanel({
-  searchQuery,
-  onSearchChange,
-  sortOption,
-  onSortChange,
-  filters,
-  onFilterChange,
-  onClearAll,
-  activeTab,
-  onTabChange,
-  projects,
-}: ProjectConditionsPanelProps) {
-  const [internalTab, setInternalTab] = useState<'sort' | 'filter'>('sort');
-  const [activeFilterTab, setActiveFilterTab] = useState(0);
-  const mainTab = activeTab ?? internalTab;
-  const setMainTab = (tab: 'sort' | 'filter') => {
-    setInternalTab(tab);
-    onTabChange?.(tab);
-  };
-
-  // ステータス件数（元データから計算）
-  const statusCounts = useMemo(() => ({
-    all_met: projects.filter(p => p.evaluationStatus === 'all_met').length,
-    other_only_unmet: projects.filter(p => p.evaluationStatus === 'other_only_unmet').length,
-    unmet: projects.filter(p => p.evaluationStatus === 'unmet').length,
-  }), [projects]);
-
-  // フィルター件数
-  const filterCounts = {
-    workStatus: filters.workStatuses.length,
-    evaluationStatus: filters.evaluationStatuses.length,
-    priority: filters.priorities.length,
-    bidType: filters.bidTypes.length,
-    category: filters.categories.length,
-    prefecture: filters.prefectures.length,
-    organization: filters.organizations.length,
-  };
-  const totalFilterCount = Object.values(filterCounts).reduce((a, b) => a + b, 0);
-  const hasConditions = searchQuery.trim() || sortOption !== null || totalFilterCount > 0;
-
-  // トグル関数
-  const toggleWorkStatus = (status: WorkStatus) => {
-    const newStatuses = filters.workStatuses.includes(status)
-      ? filters.workStatuses.filter(s => s !== status)
-      : [...filters.workStatuses, status];
-    onFilterChange({ ...filters, workStatuses: newStatuses });
-  };
-
-  const toggleEvaluationStatus = (status: EvaluationStatus) => {
-    const newStatuses = filters.evaluationStatuses.includes(status)
-      ? filters.evaluationStatuses.filter(s => s !== status)
-      : [...filters.evaluationStatuses, status];
-    onFilterChange({ ...filters, evaluationStatuses: newStatuses });
-  };
-
-  const togglePriority = (priority: CompanyPriority) => {
-    const newPriorities = filters.priorities.includes(priority)
-      ? filters.priorities.filter(p => p !== priority)
-      : [...filters.priorities, priority];
-    onFilterChange({ ...filters, priorities: newPriorities });
-  };
-
-  const toggleBidType = (bidType: string) => {
-    const newBidTypes = filters.bidTypes.includes(bidType)
-      ? filters.bidTypes.filter(b => b !== bidType)
-      : [...filters.bidTypes, bidType];
-    onFilterChange({ ...filters, bidTypes: newBidTypes });
-  };
-
-  const toggleCategory = (category: string) => {
-    const newCategories = filters.categories.includes(category)
-      ? filters.categories.filter(c => c !== category)
-      : [...filters.categories, category];
-    onFilterChange({ ...filters, categories: newCategories });
-  };
-
-  const togglePrefecture = (pref: string) => {
-    const newPrefs = filters.prefectures.includes(pref)
-      ? filters.prefectures.filter(p => p !== pref)
-      : [...filters.prefectures, pref];
-    onFilterChange({ ...filters, prefectures: newPrefs });
-  };
-
-  const togglePrefRegion = (regionItems: readonly string[]) => {
-    const allSelected = regionItems.every(item => filters.prefectures.includes(item));
-    if (allSelected) {
-      onFilterChange({
-        ...filters,
-        prefectures: filters.prefectures.filter(p => !regionItems.includes(p)),
-      });
-    } else {
-      const newPrefs = new Set([...filters.prefectures, ...regionItems]);
-      onFilterChange({ ...filters, prefectures: Array.from(newPrefs) });
-    }
-  };
-
-  const getPrefRegionSelectionState = (regionItems: readonly string[]): 'all' | 'partial' | 'none' => {
-    const selectedCount = regionItems.filter(item => filters.prefectures.includes(item)).length;
-    if (selectedCount === 0) return 'none';
-    if (selectedCount === regionItems.length) return 'all';
-    return 'partial';
-  };
-
-  const toggleOrganization = (org: string) => {
-    const newOrgs = filters.organizations.includes(org)
-      ? filters.organizations.filter(o => o !== org)
-      : [...filters.organizations, org];
-    onFilterChange({ ...filters, organizations: newOrgs });
-  };
-
-  const toggleOrgRegion = (regionItems: string[]) => {
-    const allSelected = regionItems.every(item => filters.organizations.includes(item));
-    if (allSelected) {
-      onFilterChange({
-        ...filters,
-        organizations: filters.organizations.filter(o => !regionItems.includes(o)),
-      });
-    } else {
-      const newOrgs = new Set([...filters.organizations, ...regionItems]);
-      onFilterChange({ ...filters, organizations: Array.from(newOrgs) });
-    }
-  };
-
-  const getOrgRegionSelectionState = (regionItems: string[]): 'all' | 'partial' | 'none' => {
-    const selectedCount = regionItems.filter(item => filters.organizations.includes(item)).length;
-    if (selectedCount === 0) return 'none';
-    if (selectedCount === regionItems.length) return 'all';
-    return 'partial';
-  };
-
-  const sectionDivider = {
-    borderBottom: `1px solid ${rightPanelColors.border}`,
-    pb: 2.5,
-    mb: 2.5,
-  };
-
-  // フィルターコンテンツをレンダリング
-  const renderFilterContent = () => {
-    const tabId = FILTER_TABS[activeFilterTab].id;
-
-    switch (tabId) {
-      case 'workStatus':
-        return (
-          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-            {WORK_STATUS_OPTIONS.map((status) => {
-              const config = workStatusConfig[status];
-              return (
-                <FilterButton
-                  key={status}
-                  label={config.label}
-                  selected={filters.workStatuses.includes(status)}
-                  onClick={() => toggleWorkStatus(status)}
-                  color={config.color}
-                  bgColor={config.bgColor}
-                />
-              );
-            })}
-          </Box>
-        );
-
-      case 'evaluationStatus':
-        return (
-          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-            {EVALUATION_STATUS_OPTIONS.map((status) => {
-              const config = evaluationStatusConfig[status];
-              return (
-                <FilterButton
-                  key={status}
-                  label={`${config.label} (${statusCounts[status]})`}
-                  selected={filters.evaluationStatuses.includes(status)}
-                  onClick={() => toggleEvaluationStatus(status)}
-                  color={config.color}
-                  bgColor={config.bgColor}
-                />
-              );
-            })}
-          </Box>
-        );
-
-      case 'priority':
-        return (
-          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-            {PRIORITY_OPTIONS.map((priority) => {
-              const pColor = priorityColors[priority];
-              return (
-                <FilterButton
-                  key={priority}
-                  label={priorityLabels[priority]}
-                  selected={filters.priorities.includes(priority)}
-                  onClick={() => togglePriority(priority)}
-                  color={pColor.color}
-                  bgColor={pColor.bgColor}
-                />
-              );
-            })}
-          </Box>
-        );
-
-      case 'bidType':
-        return (
-          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-            {bidTypes.map((bidType: BidType) => {
-              const config = bidTypeConfig[bidType];
-              return (
-                <FilterButton
-                  key={bidType}
-                  label={config.label}
-                  selected={filters.bidTypes.includes(bidType)}
-                  onClick={() => toggleBidType(bidType)}
-                  color={config.color}
-                  bgColor={config.bgColor}
-                />
-              );
-            })}
-          </Box>
-        );
-
-      case 'category':
-        return (
-          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-            {categories.map((cat) => (
-              <FilterButton
-                key={cat}
-                label={cat}
-                selected={filters.categories.includes(cat)}
-                onClick={() => toggleCategory(cat)}
-              />
-            ))}
-          </Box>
-        );
-
-      case 'prefecture':
-        return (
-          <Box>
-            {prefecturesByRegion.map((region) => {
-              const selectionState = getPrefRegionSelectionState(region.prefectures);
-              return (
-                <Box key={region.region} sx={{ mb: 2 }}>
-                  <button
-                    onClick={() => togglePrefRegion(region.prefectures)}
-                    style={{
-                      display: 'inline-flex',
-                      alignItems: 'center',
-                      gap: '6px',
-                      fontSize: fontSizes.xs,
-                      fontWeight: 600,
-                      color: selectionState === 'none' ? rightPanelColors.textMuted : rightPanelColors.text,
-                      marginBottom: '8px',
-                      padding: '4px 10px',
-                      borderRadius: borderRadius.xs,
-                      border: 'none',
-                      cursor: 'pointer',
-                      background:
-                        selectionState === 'all'
-                          ? `${colors.accent.blue}40`
-                          : selectionState === 'partial'
-                            ? `${colors.accent.blue}26`
-                            : 'rgba(255, 255, 255, 0.08)',
-                    }}
-                  >
-                    <span
-                      style={{
-                        width: 14,
-                        height: 14,
-                        borderRadius: '3px',
-                        border: selectionState === 'none' ? `2px solid ${colors.text.muted}` : 'none',
-                        background:
-                          selectionState === 'all'
-                            ? colors.accent.blue
-                            : selectionState === 'partial'
-                              ? colors.status.info.light
-                              : 'transparent',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        fontSize: '10px',
-                        color: colors.text.white,
-                      }}
-                    >
-                      {selectionState === 'all' && '✓'}
-                      {selectionState === 'partial' && '−'}
-                    </span>
-                    {region.region}
-                    <span style={{ fontSize: fontSizes.xs, color: rightPanelColors.textMuted }}>
-                      ({region.prefectures.filter(item => filters.prefectures.includes(item)).length}/{region.prefectures.length})
-                    </span>
-                  </button>
-                  <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-                    {region.prefectures.map((pref) => (
-                      <FilterButton
-                        key={pref}
-                        label={pref}
-                        selected={filters.prefectures.includes(pref)}
-                        onClick={() => togglePrefecture(pref)}
-                      />
-                    ))}
-                  </Box>
-                </Box>
-              );
-            })}
-          </Box>
-        );
-
-      case 'organization':
-        return (
-          <Box>
-            {organizationGroupsByRegion.map((group) => {
-              const selectionState = getOrgRegionSelectionState(group.items);
-              return (
-                <Box key={group.region} sx={{ mb: 2 }}>
-                  <button
-                    onClick={() => toggleOrgRegion(group.items)}
-                    style={{
-                      display: 'inline-flex',
-                      alignItems: 'center',
-                      gap: '6px',
-                      fontSize: fontSizes.xs,
-                      fontWeight: 600,
-                      color: selectionState === 'none' ? rightPanelColors.textMuted : rightPanelColors.text,
-                      marginBottom: '8px',
-                      padding: '4px 10px',
-                      borderRadius: borderRadius.xs,
-                      border: 'none',
-                      cursor: 'pointer',
-                      background:
-                        selectionState === 'all'
-                          ? `${colors.accent.blue}40`
-                          : selectionState === 'partial'
-                            ? `${colors.accent.blue}26`
-                            : 'rgba(255, 255, 255, 0.08)',
-                    }}
-                  >
-                    <span
-                      style={{
-                        width: 14,
-                        height: 14,
-                        borderRadius: '3px',
-                        border: selectionState === 'none' ? `2px solid ${colors.text.muted}` : 'none',
-                        background:
-                          selectionState === 'all'
-                            ? colors.accent.blue
-                            : selectionState === 'partial'
-                              ? colors.status.info.light
-                              : 'transparent',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        fontSize: '10px',
-                        color: colors.text.white,
-                      }}
-                    >
-                      {selectionState === 'all' && '✓'}
-                      {selectionState === 'partial' && '−'}
-                    </span>
-                    {group.region}
-                    <span style={{ fontSize: fontSizes.xs, color: rightPanelColors.textMuted }}>
-                      ({group.items.filter(item => filters.organizations.includes(item)).length}/{group.items.length})
-                    </span>
-                  </button>
-                  <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
-                    {group.items.map((org) => (
-                      <FilterButton
-                        key={org}
-                        label={org}
-                        selected={filters.organizations.includes(org)}
-                        onClick={() => toggleOrganization(org)}
-                      />
-                    ))}
-                  </Box>
-                </Box>
-              );
-            })}
-          </Box>
-        );
-
-      default:
-        return null;
-    }
-  };
-
-  return (
-    <Box sx={{ display: 'flex', flexDirection: 'column' }}>
-      {/* クリアボタン */}
-      <Box sx={sectionDivider}>
-        <Button
-          onClick={onClearAll}
-          variant="outlined"
-          size="small"
-          fullWidth
-          disabled={!hasConditions}
-          sx={{
-            color: hasConditions ? colors.accent.red : rightPanelColors.textMuted,
-            borderColor: hasConditions ? `${colors.accent.red}80` : rightPanelColors.inputBorder,
-            fontSize: fontSizes.xs,
-            py: 0.75,
-            '&:hover': {
-              borderColor: colors.accent.red,
-              backgroundColor: `${colors.accent.red}1a`,
-            },
-            '&.Mui-disabled': {
-              color: rightPanelColors.textMuted,
-              borderColor: rightPanelColors.inputBorder,
-            },
-          }}
-        >
-          すべてクリア
-        </Button>
-      </Box>
-
-      {/* 検索 */}
-      <Box sx={sectionDivider}>
-        <TextField
-          placeholder="検索..."
-          value={searchQuery}
-          onChange={onSearchChange}
-          size="small"
-          fullWidth
-          sx={{
-            '& .MuiOutlinedInput-root': {
-              fontSize: fontSizes.sm,
-              color: rightPanelColors.text,
-              backgroundColor: rightPanelColors.inputBg,
-              '& fieldset': { borderColor: rightPanelColors.inputBorder },
-              '&:hover fieldset': { borderColor: `${colors.text.light}99` },
-              '&.Mui-focused fieldset': { borderColor: colors.accent.blue },
-            },
-            '& .MuiInputBase-input::placeholder': { color: rightPanelColors.textMuted },
-          }}
-          InputProps={{
-            startAdornment: (
-              <InputAdornment position="start">
-                <SearchIcon sx={{ ...iconStyles.medium, color: rightPanelColors.textMuted }} />
-              </InputAdornment>
-            ),
-            endAdornment: searchQuery && (
-              <InputAdornment position="end">
-                <IconButton
-                  size="small"
-                  onClick={() => onSearchChange({ target: { value: '' } } as React.ChangeEvent<HTMLInputElement>)}
-                  sx={{ color: rightPanelColors.textMuted }}
-                >
-                  <CloseIcon fontSize="small" />
-                </IconButton>
-              </InputAdornment>
-            ),
-          }}
-        />
-      </Box>
-
-      {/* ソート/フィルター タブ */}
-      <Box>
-        <Box sx={{ display: 'flex', borderBottom: `1px solid ${rightPanelColors.border}`, mb: 2 }}>
-          <Box
-            onClick={() => setMainTab('sort')}
-            sx={{
-              flex: 1,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              gap: 0.75,
-              py: 1.25,
-              cursor: 'pointer',
-              borderBottom: mainTab === 'sort' ? `2px solid ${colors.accent.blue}` : '2px solid transparent',
-              color: mainTab === 'sort' ? rightPanelColors.text : rightPanelColors.textMuted,
-              fontWeight: 600,
-              fontSize: fontSizes.sm,
-            }}
-          >
-            <SortIcon sx={iconStyles.medium} />
-            ソート
-          </Box>
-          <Box
-            onClick={() => setMainTab('filter')}
-            sx={{
-              flex: 1,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              gap: 0.75,
-              py: 1.25,
-              cursor: 'pointer',
-              borderBottom: mainTab === 'filter' ? `2px solid ${colors.accent.blue}` : '2px solid transparent',
-              color: mainTab === 'filter' ? rightPanelColors.text : rightPanelColors.textMuted,
-              fontWeight: 600,
-              fontSize: fontSizes.sm,
-            }}
-          >
-            <FilterIcon sx={iconStyles.medium} />
-            フィルター
-            {totalFilterCount > 0 && (
-              <Box
-                component="span"
-                sx={{
-                  padding: '2px 6px',
-                  borderRadius: borderRadius.xs,
-                  fontSize: fontSizes.xs,
-                  fontWeight: 600,
-                  backgroundColor: `${colors.accent.blue}4d`,
-                  color: colors.accent.blue,
-                }}
-              >
-                {totalFilterCount}
-              </Box>
-            )}
-          </Box>
-        </Box>
-
-        {/* ソートコンテンツ */}
-        {mainTab === 'sort' && (
-          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-            {SORT_FIELDS.map(({ field, label, ascLabel, descLabel }) => {
-              const ascOption = `${field}_asc` as SortOption;
-              const descOption = `${field}_desc` as SortOption;
-              const isAsc = sortOption === ascOption;
-              const isDesc = sortOption === descOption;
-              const isActive = isAsc || isDesc;
-              const buttonText = !isActive
-                ? label
-                : isAsc
-                  ? `${label}：${ascLabel}↑`
-                  : `${label}：${descLabel}↓`;
-
-              return (
-                <Box
-                  component="button"
-                  key={field}
-                  onClick={() => {
-                    if (!isActive) {
-                      onSortChange(ascOption);
-                    } else if (isAsc) {
-                      onSortChange(descOption);
-                    } else {
-                      onSortChange(null);
-                    }
-                  }}
-                  sx={{
-                    position: 'relative',
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'flex-start',
-                    width: '100%',
-                    px: 2,
-                    pl: isActive ? 2.5 : 2,
-                    py: 1.5,
-                    fontSize: fontSizes.md,
-                    fontWeight: isActive ? 600 : 500,
-                    borderRadius: borderRadius.xs,
-                    border: `1px solid ${isActive ? colors.accent.blue : rightPanelColors.inputBorder}`,
-                    cursor: 'pointer',
-                    backgroundColor: isActive ? `${colors.accent.blue}26` : 'transparent',
-                    color: isActive ? colors.text.white : rightPanelColors.textMuted,
-                    ...(isActive && {
-                      '&::before': {
-                        content: '""',
-                        position: 'absolute',
-                        left: 0,
-                        top: 0,
-                        bottom: 0,
-                        width: '3px',
-                        backgroundColor: colors.accent.blue,
-                      },
-                    }),
-                    '&:hover': {
-                      backgroundColor: isActive ? `${colors.accent.blue}33` : 'rgba(255, 255, 255, 0.06)',
-                      color: isActive ? colors.text.white : rightPanelColors.text,
-                    },
-                  }}
-                >
-                  {buttonText}
-                </Box>
-              );
-            })}
-          </Box>
-        )}
-
-        {/* フィルターコンテンツ */}
-        {mainTab === 'filter' && (
-          <Box>
-            {/* カテゴリ選択ボタン（グリッド） */}
-            <Box sx={{
-              display: 'grid',
-              gridTemplateColumns: '1fr 1fr',
-              gap: 0.75,
-              mb: 2,
-              pb: 2,
-              borderBottom: `1px solid ${rightPanelColors.border}`
-            }}>
-              {FILTER_TABS.map((tab, index) => {
-                const count = filterCounts[tab.id as keyof typeof filterCounts];
-                const isActive = activeFilterTab === index;
-                const handleDoubleClick = () => {
-                  switch (tab.id) {
-                    case 'evaluationStatus':
-                      onFilterChange({ ...filters, evaluationStatuses: [] });
-                      break;
-                    case 'priority':
-                      onFilterChange({ ...filters, priorities: [] });
-                      break;
-                    case 'workStatus':
-                      onFilterChange({ ...filters, workStatuses: [] });
-                      break;
-                    case 'bidType':
-                      onFilterChange({ ...filters, bidTypes: [] });
-                      break;
-                    case 'category':
-                      onFilterChange({ ...filters, categories: [] });
-                      break;
-                    case 'prefecture':
-                      onFilterChange({ ...filters, prefectures: [] });
-                      break;
-                    case 'organization':
-                      onFilterChange({ ...filters, organizations: [] });
-                      break;
-                  }
-                };
-                return (
-                  <Box
-                    component="button"
-                    key={tab.id}
-                    onClick={() => setActiveFilterTab(index)}
-                    onDoubleClick={handleDoubleClick}
-                    sx={{
-                      display: 'flex',
-                      flexDirection: 'column',
-                      alignItems: 'center',
-                      gap: 0.5,
-                      px: 1,
-                      py: 1.25,
-                      fontSize: fontSizes.sm,
-                      fontWeight: isActive ? 600 : 500,
-                      borderRadius: borderRadius.xs,
-                      border: `1px solid ${isActive ? colors.accent.blue : rightPanelColors.inputBorder}`,
-                      cursor: 'pointer',
-                      position: 'relative',
-                      overflow: 'hidden',
-                      backgroundColor: isActive ? `${colors.accent.blue}26` : 'transparent',
-                      color: isActive ? colors.text.white : rightPanelColors.textMuted,
-                      transition: 'all 0.2s ease',
-                      ...(isActive && {
-                        '&::before': {
-                          content: '""',
-                          position: 'absolute',
-                          left: 0,
-                          top: 0,
-                          bottom: 0,
-                          width: '3px',
-                          backgroundColor: colors.accent.blue,
-                        },
-                      }),
-                      '&:hover': {
-                        backgroundColor: isActive ? `${colors.accent.blue}33` : 'rgba(255, 255, 255, 0.06)',
-                        color: isActive ? colors.text.white : rightPanelColors.text,
-                        borderColor: isActive ? colors.accent.blue : `${colors.text.light}99`,
-                      },
-                    }}
-                  >
-                    {tab.label}
-                    {count > 0 && (
-                      <Box
-                        component="span"
-                        sx={{
-                          position: 'absolute',
-                          top: 4,
-                          right: 4,
-                          padding: '1px 5px',
-                          borderRadius: borderRadius.xs,
-                          fontSize: fontSizes.xs,
-                          fontWeight: 700,
-                          backgroundColor: isActive ? 'rgba(255,255,255,0.3)' : `${colors.accent.blue}cc`,
-                          color: colors.text.white,
-                          minWidth: 14,
-                          textAlign: 'center',
-                        }}
-                      >
-                        {count}
-                      </Box>
-                    )}
-                  </Box>
-                );
-              })}
-            </Box>
-
-            {/* フィルター詳細コンテンツ */}
-            <Box sx={{ minHeight: 100 }}>
-              {renderFilterContent()}
-            </Box>
-          </Box>
-        )}
-      </Box>
-    </Box>
-  );
-}
-
-// ナビゲーション追跡用
-const NAV_TRACKING_KEY = 'lastVisitedPath';
+import {
+  PartnerDetailHeader,
+  ProjectConditionsPanel,
+} from './partner-detail';
+import type { SortOption, ProjectFilterState } from './partner-detail';
+import {
+  WORK_STATUS_ORDER,
+  EVALUATION_STATUS_ORDER,
+  NAV_TRACKING_KEY,
+} from './partner-detail';
 
 export default function PartnerDetailPage() {
   const { id } = useParams<{ id: string }>();
@@ -814,19 +34,19 @@ export default function PartnerDetailPage() {
   const { rightPanelOpen, toggleRightPanel, closeRightPanel, isMobile } = useSidebar();
   const [conditionTab, setConditionTab] = useState<'sort' | 'filter'>('sort');
 
-  // APIからデータ取得
+  // API data fetch
   const [partner, setPartner] = useState<PartnerDetail | null>(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
 
-  // 詳細ページのパスを保存（一覧に戻った時のページ復元用）
+  // Save detail page path for page restoration when returning to list
   useEffect(() => {
     try {
       sessionStorage.setItem(NAV_TRACKING_KEY, location.pathname);
     } catch { /* ignore */ }
   }, [location.pathname]);
 
-  // パートナー情報をAPIから取得
+  // Fetch partner info from API
   useEffect(() => {
     const fetchPartner = async () => {
       if (!id) return;
@@ -848,12 +68,12 @@ export default function PartnerDetailPage() {
     fetchPartner();
   }, [id]);
 
-  // 対応案件のページネーション・検索状態
+  // Project pagination & search state
   const [projectSearchQuery, setProjectSearchQuery] = useState('');
   const [projectPage, setProjectPage] = useState(0);
   const projectPageSize = 25;
 
-  // ソート・フィルター状態
+  // Sort & filter state
   const [sortOption, setSortOption] = useState<SortOption | null>(null);
   const [filters, setFilters] = useState<ProjectFilterState>({
     workStatuses: [],
@@ -865,7 +85,7 @@ export default function PartnerDetailPage() {
     organizations: [],
   });
 
-  // サイドパネル制御
+  // Side panel control
   const handleOpenWithTab = useCallback((tab: 'sort' | 'filter') => {
     setConditionTab(tab);
     if (!rightPanelOpen) {
@@ -873,13 +93,13 @@ export default function PartnerDetailPage() {
     }
   }, [rightPanelOpen, toggleRightPanel]);
 
-  // 対応案件のフィルタリング
+  // Filter projects
   const filteredProjects = useMemo((): PastProject[] => {
     if (!partner) return [];
     const pastProjects = partner.pastProjects || [];
     if (!Array.isArray(pastProjects)) return [];
 
-    // 検索フィルター
+    // Search filter
     let filtered = pastProjects;
     if (projectSearchQuery) {
       const query = projectSearchQuery.toLowerCase();
@@ -892,7 +112,7 @@ export default function PartnerDetailPage() {
       );
     }
 
-    // フィルター適用
+    // Apply filters
     filtered = filtered.filter((p: PastProject) => {
       if (filters.workStatuses.length > 0 && !filters.workStatuses.includes(p.workStatus)) return false;
       if (filters.evaluationStatuses.length > 0 && !filters.evaluationStatuses.includes(p.evaluationStatus)) return false;
@@ -906,7 +126,7 @@ export default function PartnerDetailPage() {
       return true;
     });
 
-    // ソート適用
+    // Apply sort
     if (!sortOption) return filtered;
 
     return [...filtered].sort((a, b) => {
@@ -953,13 +173,13 @@ export default function PartnerDetailPage() {
     });
   }, [partner, projectSearchQuery, filters, sortOption]);
 
-  // ページネーションされた対応案件
+  // Paginated projects
   const paginatedProjects = useMemo(() => {
     const start = projectPage * projectPageSize;
     return filteredProjects.slice(start, start + projectPageSize);
   }, [filteredProjects, projectPage]);
 
-  // 検索変更ハンドラ
+  // Search change handler
   const handleSearchChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
     setProjectSearchQuery(e.target.value);
     setProjectPage(0);
@@ -991,7 +211,7 @@ export default function PartnerDetailPage() {
     );
   }
 
-  // データ構造の保証（デフォルト値設定）
+  // Ensure data structure defaults
   const safePartner = {
     ...partner,
     branches: partner.branches || [],
@@ -1003,49 +223,17 @@ export default function PartnerDetailPage() {
   return (
     <Box sx={{ height: '100vh', bgcolor: colors.page.background, display: 'flex', flexDirection: 'column', overflow: 'hidden' }}>
       <Box sx={{ ...pageStyles.contentArea, maxWidth: '100%', py: 3, display: 'flex', flexDirection: 'column', flex: 1, minHeight: 0 }}>
-        {/* ヘッダー */}
-        <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'stretch', mb: 2 }}>
-          <Box sx={{ pl: 2, borderLeft: '4px solid', borderColor: colors.accent.blue }}>
-            <Box sx={{ display: 'flex', alignItems: 'center', gap: 2.5, mb: 0.5, flexWrap: 'wrap' }}>
-              <Button
-                size="small"
-                startIcon={<ArrowBackIcon sx={iconStyles.small} />}
-                onClick={() => navigate('/partners')}
-                sx={{
-                  color: colors.text.muted,
-                  fontWeight: 500,
-                  fontSize: fontSizes.sm,
-                  textTransform: 'none',
-                  px: 1,
-                  py: 0.25,
-                  minWidth: 'auto',
-                  '&:hover': { backgroundColor: colors.border.light },
-                }}
-              >
-                一覧に戻る
-              </Button>
-            </Box>
-            <Typography sx={pageStyles.detailPageTitle}>
-              {safePartner.name}
-            </Typography>
-          </Box>
-          <Box sx={{ textAlign: 'right', flexShrink: 0, ml: 3, display: 'flex', flexDirection: 'column', justifyContent: 'space-between' }}>
-            <Typography sx={{ fontSize: fontSizes.md, color: colors.text.light }}>
-              No. {String(safePartner.no).padStart(8, '0')}
-            </Typography>
-            <Typography sx={{ fontSize: fontSizes.xl, fontWeight: 700, color: colors.accent.blue }}>
-              <Typography component="span" sx={{ fontSize: fontSizes.sm, fontWeight: 500, color: colors.text.muted }}>
-                対応案件実績{" "}
-              </Typography>
-              {safePartner.pastProjects.length}件
-            </Typography>
-          </Box>
-        </Box>
+        {/* Header */}
+        <PartnerDetailHeader
+          partner={safePartner}
+          projectCount={safePartner.pastProjects.length}
+          onBack={() => navigate('/partners')}
+        />
 
-        {/* メインカード + サイドパネル */}
+        {/* Main card + side panel */}
         <Box sx={{ flex: 1, display: 'flex', minHeight: 0 }}>
           <Paper sx={{ borderRadius: borderRadius.xs, overflow: 'hidden', boxShadow: '0 2px 8px rgba(0, 0, 0, 0.1)', flex: 1, display: 'flex', flexDirection: 'column' }}>
-            {/* タブ */}
+            {/* Tabs */}
             <Box sx={{ borderBottom: `1px solid ${colors.border.main}`, backgroundColor: colors.text.white }}>
               <Tabs
                 value={activeTab}
@@ -1061,36 +249,31 @@ export default function PartnerDetailPage() {
               </Tabs>
             </Box>
 
-            {/* タブコンテンツ */}
+            {/* Tab content */}
             <Box sx={{ flex: 1, overflow: 'auto', backgroundColor: colors.background.hover, display: 'flex', flexDirection: 'column' }}>
-            {/* 基本情報タブ */}
-            {activeTab === 0 && (
-              <Box sx={{ p: 2.5 }}>
-                <Box sx={{ display: 'grid', gridTemplateColumns: { xs: '1fr', md: '1fr 1fr' }, gap: 2.5 }}>
-                  {/* 左カラム */}
-                  <PartnerBasicInfo partner={safePartner} />
-
-                  {/* 右カラム */}
-                  <PartnerQualifications partner={safePartner} />
+              {activeTab === 0 && (
+                <Box sx={{ p: 2.5 }}>
+                  <Box sx={{ display: 'grid', gridTemplateColumns: { xs: '1fr', md: '1fr 1fr' }, gap: 2.5 }}>
+                    <PartnerBasicInfo partner={safePartner} />
+                    <PartnerQualifications partner={safePartner} />
+                  </Box>
                 </Box>
-              </Box>
-            )}
+              )}
 
-            {/* 対応案件タブ */}
-            {activeTab === 1 && (
-              <PartnerHistory
-                projects={paginatedProjects}
-                page={projectPage}
-                pageSize={projectPageSize}
-                filteredCount={filteredProjects.length}
-                onPageChange={setProjectPage}
-                onNavigate={navigate}
-              />
-            )}
+              {activeTab === 1 && (
+                <PartnerHistory
+                  projects={paginatedProjects}
+                  page={projectPage}
+                  pageSize={projectPageSize}
+                  filteredCount={filteredProjects.length}
+                  onPageChange={setProjectPage}
+                  onNavigate={navigate}
+                />
+              )}
             </Box>
           </Paper>
 
-          {/* 右サイドパネル（対応案件タブの時のみ表示） */}
+          {/* Right side panel (visible only on project history tab) */}
           {activeTab === 1 && (
             <RightSidePanel
               open={rightPanelOpen}
