+++ judgesystem/frontend/app/src/pages/orderer-detail/AnnouncementConditionsPanel.tsx
@@ -1,0 +1,537 @@
+import { useState } from 'react';
+import {
+  Box,
+  Button,
+  TextField,
+  InputAdornment,
+  IconButton,
+} from '@mui/material';
+import {
+  Search as SearchIcon,
+  SwapVert as SortIcon,
+  FilterAlt as FilterIcon,
+  Close as CloseIcon,
+} from '@mui/icons-material';
+import { announcementStatusConfig, bidTypeConfig } from '../../data';
+import { categories } from '../../constants/categories';
+import { bidTypes } from '../../constants/bidType';
+import { prefecturesByRegion } from '../../constants/prefectures';
+import type { BidType } from '../../types/announcement';
+import { colors, fontSizes, borderRadius, rightPanelColors, iconStyles } from '../../constants/styles';
+import { FilterButton } from '../../components/common';
+import type {
+  SortOption,
+  StatusFilter,
+  AnnouncementFilterState,
+} from './types';
+import {
+  SORT_FIELDS,
+  FILTER_TABS,
+  STATUS_FILTERS,
+} from './types';
+
+/** Props for AnnouncementConditionsPanel */
+export interface AnnouncementConditionsPanelProps {
+  searchQuery: string;
+  onSearchChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  sortOption: SortOption | null;
+  onSortChange: (option: SortOption | null) => void;
+  filters: AnnouncementFilterState;
+  onFilterChange: (filters: AnnouncementFilterState) => void;
+  onClearAll: () => void;
+  activeTab?: 'sort' | 'filter';
+  onTabChange?: (tab: 'sort' | 'filter') => void;
+}
+
+/**
+ * Side panel for controlling search, sort, and filter conditions
+ * on the orderer's announcements list.
+ */
+export function AnnouncementConditionsPanel({
+  searchQuery,
+  onSearchChange,
+  sortOption,
+  onSortChange,
+  filters,
+  onFilterChange,
+  onClearAll,
+  activeTab,
+  onTabChange,
+}: AnnouncementConditionsPanelProps) {
+  const [internalTab, setInternalTab] = useState<'sort' | 'filter'>('sort');
+  const [activeFilterTab, setActiveFilterTab] = useState(0);
+  const mainTab = activeTab ?? internalTab;
+  const setMainTab = (tab: 'sort' | 'filter') => {
+    setInternalTab(tab);
+    onTabChange?.(tab);
+  };
+
+  // Filter counts
+  const filterCounts = {
+    status: filters.statuses.length,
+    bidType: filters.bidTypes.length,
+    category: filters.categories.length,
+    prefecture: filters.prefectures.length,
+  };
+  const totalFilterCount = Object.values(filterCounts).reduce((a, b) => a + b, 0);
+  const hasConditions = searchQuery.trim() || sortOption !== null || totalFilterCount > 0;
+
+  // Toggle functions
+  const toggleStatus = (status: StatusFilter) => {
+    const newStatuses = filters.statuses.includes(status)
+      ? filters.statuses.filter(s => s !== status)
+      : [...filters.statuses, status];
+    onFilterChange({ ...filters, statuses: newStatuses });
+  };
+
+  const toggleBidType = (bidType: string) => {
+    const newBidTypes = filters.bidTypes.includes(bidType)
+      ? filters.bidTypes.filter(b => b !== bidType)
+      : [...filters.bidTypes, bidType];
+    onFilterChange({ ...filters, bidTypes: newBidTypes });
+  };
+
+  const toggleCategory = (category: string) => {
+    const newCategories = filters.categories.includes(category)
+      ? filters.categories.filter(c => c !== category)
+      : [...filters.categories, category];
+    onFilterChange({ ...filters, categories: newCategories });
+  };
+
+  const togglePrefecture = (pref: string) => {
+    const newPrefs = filters.prefectures.includes(pref)
+      ? filters.prefectures.filter(p => p !== pref)
+      : [...filters.prefectures, pref];
+    onFilterChange({ ...filters, prefectures: newPrefs });
+  };
+
+  // Region group toggle (prefectures)
+  const togglePrefRegion = (regionItems: readonly string[]) => {
+    const allSelected = regionItems.every(item => filters.prefectures.includes(item));
+    if (allSelected) {
+      onFilterChange({
+        ...filters,
+        prefectures: filters.prefectures.filter(p => !regionItems.includes(p)),
+      });
+    } else {
+      const newPrefs = new Set([...filters.prefectures, ...regionItems]);
+      onFilterChange({ ...filters, prefectures: Array.from(newPrefs) });
+    }
+  };
+
+  const getPrefRegionSelectionState = (regionItems: readonly string[]): 'all' | 'partial' | 'none' => {
+    const selectedCount = regionItems.filter(item => filters.prefectures.includes(item)).length;
+    if (selectedCount === 0) return 'none';
+    if (selectedCount === regionItems.length) return 'all';
+    return 'partial';
+  };
+
+  const sectionDivider = {
+    borderBottom: `1px solid ${rightPanelColors.border}`,
+    pb: 2.5,
+    mb: 2.5,
+  };
+
+  // Render filter content by active tab
+  const renderFilterContent = () => {
+    const tabId = FILTER_TABS[activeFilterTab].id;
+
+    switch (tabId) {
+      case 'status':
+        return (
+          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
+            {STATUS_FILTERS.map((status) => {
+              const config = announcementStatusConfig[status];
+              return (
+                <FilterButton
+                  key={status}
+                  label={config.label}
+                  selected={filters.statuses.includes(status)}
+                  onClick={() => toggleStatus(status)}
+                  color={config.color}
+                  bgColor={config.bgColor}
+                />
+              );
+            })}
+          </Box>
+        );
+
+      case 'bidType':
+        return (
+          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
+            {bidTypes.map((bidType: BidType) => {
+              const config = bidTypeConfig[bidType];
+              return (
+                <FilterButton
+                  key={bidType}
+                  label={config.label}
+                  selected={filters.bidTypes.includes(bidType)}
+                  onClick={() => toggleBidType(bidType)}
+                  color={config.color}
+                  bgColor={config.bgColor}
+                />
+              );
+            })}
+          </Box>
+        );
+
+      case 'category':
+        return (
+          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.75 }}>
+            {categories.map((cat) => (
+              <FilterButton
+                key={cat}
+                label={cat}
+                selected={filters.categories.includes(cat)}
+                onClick={() => toggleCategory(cat)}
+              />
+            ))}
+          </Box>
+        );
+
+      case 'prefecture':
+        return (
+          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
+            {prefecturesByRegion.map((region) => {
+              const selectionState = getPrefRegionSelectionState(region.prefectures);
+              return (
+                <Box key={region.region}>
+                  <Box
+                    component="button"
+                    onClick={() => togglePrefRegion(region.prefectures)}
+                    sx={{
+                      display: 'flex',
+                      alignItems: 'center',
+                      gap: 1,
+                      mb: 1,
+                      py: 0.5,
+                      px: 1,
+                      fontSize: fontSizes.sm,
+                      fontWeight: 600,
+                      cursor: 'pointer',
+                      border: 'none',
+                      borderRadius: borderRadius.xs,
+                      backgroundColor: selectionState !== 'none' ? `${colors.accent.blue}33` : 'transparent',
+                      color: selectionState !== 'none' ? colors.accent.blue : rightPanelColors.text,
+                      '&:hover': { backgroundColor: `${colors.accent.blue}26` },
+                    }}
+                  >
+                    {region.region}
+                    {selectionState === 'partial' && (
+                      <Box
+                        component="span"
+                        sx={{
+                          fontSize: fontSizes.xs,
+                          color: colors.accent.blue,
+                          opacity: 0.8,
+                        }}
+                      >
+                        (一部)
+                      </Box>
+                    )}
+                  </Box>
+                  <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.5 }}>
+                    {region.prefectures.map((pref) => (
+                      <FilterButton
+                        key={pref}
+                        label={pref}
+                        selected={filters.prefectures.includes(pref)}
+                        onClick={() => togglePrefecture(pref)}
+                      />
+                    ))}
+                  </Box>
+                </Box>
+              );
+            })}
+          </Box>
+        );
+
+      default:
+        return null;
+    }
+  };
+
+  return (
+    <Box sx={{ display: 'flex', flexDirection: 'column' }}>
+      {/* Clear button */}
+      <Box sx={sectionDivider}>
+        <Button
+          onClick={onClearAll}
+          variant="outlined"
+          size="small"
+          fullWidth
+          disabled={!hasConditions}
+          sx={{
+            color: hasConditions ? colors.accent.red : rightPanelColors.textMuted,
+            borderColor: hasConditions ? `${colors.accent.red}80` : rightPanelColors.inputBorder,
+            fontSize: fontSizes.xs,
+            py: 0.75,
+            '&:hover': {
+              borderColor: colors.accent.red,
+              backgroundColor: `${colors.accent.red}1a`,
+            },
+            '&.Mui-disabled': {
+              color: rightPanelColors.textMuted,
+              borderColor: rightPanelColors.inputBorder,
+            },
+          }}
+        >
+          すべてクリア
+        </Button>
+      </Box>
+
+      {/* Search */}
+      <Box sx={sectionDivider}>
+        <TextField
+          placeholder="検索..."
+          value={searchQuery}
+          onChange={onSearchChange}
+          size="small"
+          fullWidth
+          sx={{
+            '& .MuiOutlinedInput-root': {
+              fontSize: fontSizes.sm,
+              color: rightPanelColors.text,
+              backgroundColor: rightPanelColors.inputBg,
+              '& fieldset': { borderColor: rightPanelColors.inputBorder },
+              '&:hover fieldset': { borderColor: `${colors.text.light}99` },
+              '&.Mui-focused fieldset': { borderColor: colors.accent.blue },
+            },
+            '& .MuiInputBase-input::placeholder': { color: rightPanelColors.textMuted },
+          }}
+          InputProps={{
+            startAdornment: (
+              <InputAdornment position="start">
+                <SearchIcon sx={{ ...iconStyles.medium, color: rightPanelColors.textMuted }} />
+              </InputAdornment>
+            ),
+            endAdornment: searchQuery && (
+              <InputAdornment position="end">
+                <IconButton
+                  size="small"
+                  onClick={() => onSearchChange({ target: { value: '' } } as React.ChangeEvent<HTMLInputElement>)}
+                  sx={{ color: rightPanelColors.textMuted }}
+                >
+                  <CloseIcon fontSize="small" />
+                </IconButton>
+              </InputAdornment>
+            ),
+          }}
+        />
+      </Box>
+
+      {/* Sort / Filter tabs */}
+      <Box>
+        <Box sx={{ display: 'flex', borderBottom: `1px solid ${rightPanelColors.border}`, mb: 2 }}>
+          <Box
+            onClick={() => setMainTab('sort')}
+            sx={{
+              flex: 1,
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center',
+              gap: 0.75,
+              py: 1.25,
+              cursor: 'pointer',
+              borderBottom: mainTab === 'sort' ? `2px solid ${colors.accent.blue}` : '2px solid transparent',
+              color: mainTab === 'sort' ? rightPanelColors.text : rightPanelColors.textMuted,
+              fontWeight: 600,
+              fontSize: fontSizes.sm,
+            }}
+          >
+            <SortIcon sx={iconStyles.medium} />
+            ソート
+          </Box>
+          <Box
+            onClick={() => setMainTab('filter')}
+            sx={{
+              flex: 1,
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center',
+              gap: 0.75,
+              py: 1.25,
+              cursor: 'pointer',
+              borderBottom: mainTab === 'filter' ? `2px solid ${colors.accent.blue}` : '2px solid transparent',
+              color: mainTab === 'filter' ? rightPanelColors.text : rightPanelColors.textMuted,
+              fontWeight: 600,
+              fontSize: fontSizes.sm,
+            }}
+          >
+            <FilterIcon sx={iconStyles.medium} />
+            フィルター
+            {totalFilterCount > 0 && (
+              <Box
+                component="span"
+                sx={{
+                  padding: '2px 6px',
+                  borderRadius: borderRadius.xs,
+                  fontSize: fontSizes.xs,
+                  fontWeight: 600,
+                  backgroundColor: `${colors.accent.blue}4d`,
+                  color: colors.accent.blue,
+                }}
+              >
+                {totalFilterCount}
+              </Box>
+            )}
+          </Box>
+        </Box>
+
+        {/* Sort content */}
+        {mainTab === 'sort' && (
+          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
+            {SORT_FIELDS.map(({ field, label, ascLabel, descLabel }) => {
+              const ascOption = `${field}_asc` as SortOption;
+              const descOption = `${field}_desc` as SortOption;
+              const isAsc = sortOption === ascOption;
+              const isDesc = sortOption === descOption;
+              const isActive = isAsc || isDesc;
+              const buttonText = !isActive
+                ? label
+                : isAsc
+                  ? `${label}：${ascLabel}↑`
+                  : `${label}：${descLabel}↓`;
+
+              return (
+                <Box
+                  component="button"
+                  key={field}
+                  onClick={() => {
+                    if (!isActive) {
+                      onSortChange(ascOption);
+                    } else if (isAsc) {
+                      onSortChange(descOption);
+                    } else {
+                      onSortChange(null);
+                    }
+                  }}
+                  sx={{
+                    position: 'relative',
+                    display: 'flex',
+                    alignItems: 'center',
+                    justifyContent: 'flex-start',
+                    width: '100%',
+                    px: 2,
+                    pl: isActive ? 2.5 : 2,
+                    py: 1.5,
+                    fontSize: fontSizes.md,
+                    fontWeight: isActive ? 600 : 500,
+                    borderRadius: borderRadius.xs,
+                    border: `1px solid ${isActive ? colors.accent.blue : rightPanelColors.inputBorder}`,
+                    cursor: 'pointer',
+                    backgroundColor: isActive ? `${colors.accent.blue}26` : 'transparent',
+                    color: isActive ? colors.text.white : rightPanelColors.textMuted,
+                    ...(isActive && {
+                      '&::before': {
+                        content: '""',
+                        position: 'absolute',
+                        left: 0,
+                        top: 0,
+                        bottom: 0,
+                        width: '3px',
+                        backgroundColor: colors.accent.blue,
+                      },
+                    }),
+                    '&:hover': {
+                      backgroundColor: isActive ? `${colors.accent.blue}33` : 'rgba(255, 255, 255, 0.06)',
+                      color: isActive ? colors.text.white : rightPanelColors.text,
+                    },
+                  }}
+                >
+                  {buttonText}
+                </Box>
+              );
+            })}
+          </Box>
+        )}
+
+        {/* Filter content */}
+        {mainTab === 'filter' && (
+          <Box>
+            {/* Category selection buttons (grid) */}
+            <Box sx={{
+              display: 'grid',
+              gridTemplateColumns: '1fr 1fr',
+              gap: 0.75,
+              mb: 2,
+              pb: 2,
+              borderBottom: `1px solid ${rightPanelColors.border}`
+            }}>
+              {FILTER_TABS.map((tab, index) => {
+                const count = filterCounts[tab.id as keyof typeof filterCounts];
+                const isActive = activeFilterTab === index;
+                return (
+                  <Box
+                    component="button"
+                    key={tab.id}
+                    onClick={() => setActiveFilterTab(index)}
+                    sx={{
+                      display: 'flex',
+                      flexDirection: 'column',
+                      alignItems: 'center',
+                      gap: 0.5,
+                      px: 1,
+                      py: 1.25,
+                      fontSize: fontSizes.sm,
+                      fontWeight: isActive ? 600 : 500,
+                      borderRadius: borderRadius.xs,
+                      border: `1px solid ${isActive ? colors.accent.blue : rightPanelColors.inputBorder}`,
+                      cursor: 'pointer',
+                      position: 'relative',
+                      overflow: 'hidden',
+                      backgroundColor: isActive ? `${colors.accent.blue}26` : 'transparent',
+                      color: isActive ? colors.text.white : rightPanelColors.textMuted,
+                      ...(isActive && {
+                        '&::before': {
+                          content: '""',
+                          position: 'absolute',
+                          left: 0,
+                          top: 0,
+                          bottom: 0,
+                          width: '3px',
+                          backgroundColor: colors.accent.blue,
+                        },
+                      }),
+                      '&:hover': {
+                        backgroundColor: isActive ? `${colors.accent.blue}33` : 'rgba(255, 255, 255, 0.06)',
+                        color: isActive ? colors.text.white : rightPanelColors.text,
+                      },
+                    }}
+                  >
+                    {tab.label}
+                    {count > 0 && (
+                      <Box
+                        component="span"
+                        sx={{
+                          position: 'absolute',
+                          top: 4,
+                          right: 4,
+                          padding: '1px 5px',
+                          borderRadius: borderRadius.xs,
+                          fontSize: fontSizes.xs,
+                          fontWeight: 700,
+                          backgroundColor: isActive ? 'rgba(255,255,255,0.3)' : `${colors.accent.blue}cc`,
+                          color: colors.text.white,
+                          minWidth: 14,
+                          textAlign: 'center',
+                        }}
+                      >
+                        {count}
+                      </Box>
+                    )}
+                  </Box>
+                );
+              })}
+            </Box>
+
+            {/* Filter detail content */}
+            <Box sx={{ minHeight: 100 }}>
+              {renderFilterContent()}
+            </Box>
+          </Box>
+        )}
+      </Box>
+    </Box>
+  );
+}
